feat: describe service and repo interfaces for purchases

+++ internal/infrastructure/models/material.go
@@ -5,7 +5,7 @@ import "time"
 type Material struct {
 	ID        uint      `gorm:"primaryKey"`
 	Name      string    `gorm:"type:varchar(255);not null"`
-	Count     int       `gorm:"not null"`
+	Count     float64   `gorm:"not null"`
 	Measure   string    `gorm:"type:varchar(255);not null"`
 	Cost      float64   `gorm:"not null"`
 	CreatedAt time.Time `gorm:"autoCreateTime"`
